+++ firmware_17xx/synth/utils.h
@@ -30,7 +30,7 @@ uint16_t scaleU16U16(uint16_t a, uint16_t b);
 int16_t scaleU16S16(uint16_t a, int16_t b);
 
 uint16_t lerp(uint16_t a,uint16_t b,uint8_t x);
-uint16_t lerp16(uint16_t a,uint16_t b,uint32_t x);
+int32_t lerp16(uint16_t a,uint16_t b,int32_t x);
 uint16_t herp(int32_t alpha, int32_t cur, int32_t prev, int32_t prev2, int32_t prev3, int8_t frac_shift);
 void resample(const uint16_t * src, uint16_t * dst, uint16_t src_samples, uint16_t dst_samples);
 uint16_t computeShape(uint32_t phase, const uint16_t lookup[], int8_t interpolate);
